union and improve size field functions
